SEQUOIADBMAINSTREAM-2658 task-节点健康检测快照 repair拼写错误，百分比用INT32表示
git-svn-id: http://192.168.20.11/sequoiadb/trunk@33970 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/mon/monDump.cpp
@@ -545,7 +545,7 @@ namespace engine
       INT64 commitLimit       = 0 ;
       INT64 committedAS       = 0 ;
       INT64 VMLimitProc       = 0 ;
-      INT64 loadPctVM         = 0 ;
+      INT32 loadPctVM         = 0 ;
       INT64 allocatedVMProc   = 0 ;
       INT64 rss               = 0 ;
       ossProcLimits limInfo ;
